+++ idt.c
@@ -20,7 +20,7 @@ idt_ptr_t idtp;
 
 extern void idt_load();
 
-void idt_set_gate(uint8_t num, unsigned long base, uint16_t sel, uint8_t flags)
+void idt_set_gate(uint8_t num, uint32_t base, uint16_t sel, uint8_t flags)
 {
     idt[num].base_lo = base & 0xFFFF;
     idt[num].base_hi = (base >> 16) & 0xFFFF;
